Do no re-invent the wheel.
Well, it goes a bit against the initial purpose of the project
where we're re-inventing the wheel. But for the first apporach
we don't really need to re-invent the wheel, and we can use a
line by line scanner which abstracts and cleans-up the whitespace
in the lines. All given from the standard library!