Separate tokenization from AST
If we just load everything in memory this will be much faster to
ensure the AST is well built and the compiler can be informative.
Simplify for now! But the work continues... as this challenge of
allowing ANY whitespace to be a delimiter seems to be bitting us
in the a**.