Simplify language and finish AST
And then the realization came to mind, do we even need booleans?
Given that they're represented by 8bit integers, that looked like
redundancy. Remove all boolean expressions and simplify the whole
thing with only integer expressions. Finishing building the AST.

Add a roadmap as well, the objective will be to extend the compiler
with plugins to sort out all the complex stuff if a higher level
of language is needed.